+++ src/senaite/core/config/vocabularies.py
@@ -30,3 +30,22 @@ SAMPLE_PRESERVATION_CATEGORIES = (
         default=u"Lab Preservation"
     )),
 )
+
+ANALYSIS_TYPES = (
+    ("a", _(
+        u"analysis_type_analysis",
+        default=u"Analysis"
+    )),
+    ("b", _(
+        u"analysis_type_blank",
+        default=u"Blank"
+    )),
+    ("c", _(
+        u"analysis_type_control",
+        default=u"Control"
+    )),
+    ("d", _(
+        u"analysis_type_duplicate",
+        default=u"Duplicate"
+    )),
+)