+++ src/senaite/core/exportimport/instruments/panalytical/omnia/__init__.py
@@ -18,12 +18,14 @@
 # Copyright 2018-2024 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-""" Omnia Axios XRF
-"""
 from datetime import datetime
+
 from bika.lims.utils import to_unicode
-from senaite.core.exportimport.instruments.resultsimport import \
-    AnalysisResultsImporter, InstrumentCSVResultsFileParser
+from senaite.core.exportimport.instruments.importer import \
+    AnalysisResultsImporter
+from senaite.core.exportimport.instruments.parser import \
+    InstrumentCSVResultsFileParser
+
 
 class AxiosXrfCSVMultiParser(InstrumentCSVResultsFileParser):